+++ semcore/dropdown-menu/__tests__/dropdown-menu.axe-test.tsx
@@ -3,7 +3,7 @@ import { e2eStandToHtml } from '@semcore/testing-utils/e2e-stand';
 
 test.describe('Dropdown-menu', () => {
   test('Basic usage', async ({ page }) => {
-    const standPath = 'website/docs/components/dropdown-menu/examples/basic.tsx';
+    const standPath = 'stories/components/dropdown-menu/docs/examples/basic.tsx';
     const htmlContent = await e2eStandToHtml(standPath, 'en');
 
     await page.setContent(htmlContent);