+++ semcore/select/__tests__/select.browser-test.tsx
@@ -3,7 +3,7 @@ import { e2eStandToHtml } from '@semcore/testing-utils/e2e-stand';
 
 test.describe('Select', () => {
   test('Traps focus', async ({ page }) => {
-    const standPath = 'website/docs/components/select/examples/options_filtering.tsx';
+    const standPath = 'stories/components/select/docs/examples/options_filtering.tsx';
     const htmlContent = await e2eStandToHtml(standPath, 'en');
 
     await page.setContent(htmlContent);
@@ -16,7 +16,7 @@ test.describe('Select', () => {
     await expect(page).toHaveScreenshot();
   });
   test('Returns focus', async ({ page }) => {
-    const standPath = 'website/docs/components/select/examples/options_filtering.tsx';
+    const standPath = 'stories/components/select/docs/examples/options_filtering.tsx';
     const htmlContent = await e2eStandToHtml(standPath, 'en');
 
     await page.setContent(htmlContent);
@@ -34,7 +34,7 @@ test.describe('Select', () => {
     await expect(page).toHaveScreenshot();
   });
   test('Traps focus and handles clicks', async ({ page }) => {
-    const standPath = 'website/docs/components/select/examples/options_filtering.tsx';
+    const standPath = 'stories/components/select/docs/examples/options_filtering.tsx';
     const htmlContent = await e2eStandToHtml(standPath, 'en');
 
     await page.setContent(htmlContent);
@@ -62,5 +62,42 @@ test.describe('Select', () => {
     const triggerLocatior = await page.locator('[data-ui-name="Select.Trigger"]');
 
     await expect(await triggerLocatior.textContent()).toBe('Banana');
+  });
+  test('Show hint for clear input button by keyboard', async ({ page }) => {
+    const standPath = 'stories/components/select/docs/examples/options_filtering.tsx';
+    const htmlContent = await e2eStandToHtml(standPath, 'en');
+
+    await page.setContent(htmlContent);
+
+    await page.keyboard.press('Tab');
+    await page.keyboard.press('Space');
+
+    await page.waitForSelector('input');
+
+    await page.keyboard.type('Test');
+    await page.keyboard.press('Tab');
+
+    await new Promise((resolve) => setTimeout(resolve, 500));
+
+    await expect(page).toHaveScreenshot();
+  });
+
+  test('Show hint for clear input button by mouse', async ({ page }) => {
+    const standPath = 'stories/components/select/docs/examples/options_filtering.tsx';
+    const htmlContent = await e2eStandToHtml(standPath, 'en');
+
+    await page.setContent(htmlContent);
+    const trigger = await page.locator('[data-ui-name="Select.Trigger"]');
+    await trigger.click();
+
+    const input = await page.waitForSelector('input');
+    await input.click();
+    await page.keyboard.type('Test');
+
+    const close = await page.locator('[data-ui-name="InputSearch.Clear"]');
+    await close.hover();
+
+    await new Promise((resolve) => setTimeout(resolve, 500));
+    await expect(page).toHaveScreenshot();
   });
 });