+++ stories/components/button/docs/examples/button_with_icon.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import Button from 'intergalactic/button';
+import CheckM from '@semcore/icon/Check/m';
+
+const Demo = () => {
+  return (
+    <Button title='Confirm'>
+      <Button.Addon>
+        <CheckM />
+      </Button.Addon>
+    </Button>
+  );
+};
+
+export default Demo;