+++ stories/components/pagination/docs/examples/basic_usage.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import Pagination from '@semcore/pagination';
+
+const Demo = () => {
+  return <Pagination totalPages={122360} />;
+};
+
+export default Demo;