+++ stories/patterns/filters/advanced-filters/docs/index.stories.tsx
@@ -1,0 +1,13 @@
+import type { Meta, StoryObj } from '@storybook/react';
+
+import FiltersWithConditionsEx from './examples/filters-with-filter-conditions';
+
+const meta: Meta = {
+  title: 'Patterns/Filters/FiltersWithConditions',
+};
+
+export default meta;
+
+export const FiltersWithConditions: StoryObj = {
+  render: FiltersWithConditionsEx,
+};